Adjust some structs for psp2 and rename a function argument
MKIF structs, excluding SGXMKIF_HOST_CTL, should be at a size corresponding with what's expected on PSP2.

Fixed errors in PSP2 specific PVRSRV_2D_SGX_KICK.

ui32HeapSegmentSize describes the size of a segment of the heap the memory is being mapped into, and is used to avoid mapping memory in a manner which overlaps heap segments.